refactor (Finished) : step3 - Miss, Spare, Strike 상태의 동일 기능 Abstract Class로 분리
