+++ src/main/java/bowling/domian/state/Miss.java
@@ -3,8 +3,7 @@ package bowling.domian.state;
 import bowling.domian.frame.Score;
 import bowling.domian.state.exception.BowlFinishedException;
 
-public class Miss implements State {
-    private static final String BOWL_FINISHED_EXCEPTION_MSG = "두 번의 투구가 모두 끝났습니다!";
+public class Miss extends Finished {
 
     private Pins first;
     private Pins second;
@@ -12,21 +11,6 @@ public class Miss implements State {
     public Miss(Pins first, Pins second) {
         this.first = first;
         this.second = second;
-    }
-
-    @Override
-    public State bowl(int falledPinsCount) {
-        throw new BowlFinishedException(BOWL_FINISHED_EXCEPTION_MSG);
-    }
-
-    @Override
-    public boolean isFinished() {
-        return true;
-    }
-
-    @Override
-    public boolean canGetScore() {
-        return false;
     }
 
     @Override
